Minor : changed int to int64_t for Events::Set dt
libremidi ticks are of this type. This is to avoid a forced cast from a large integer to int in the ossia binding.

+++ src/core/Event.h
@@ -69,7 +69,7 @@ namespace Events {
 
 template <typename T>
 struct Set {
-    int dt;
+    int64_t dt;
     std::vector<T> events;
 
     // Used for sorting IN THE CASE OF ABSOLUTE TICKS
